#3 Refactoring
