+++ src/app/guards/role-auth.guard.ts
@@ -1,19 +1,20 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import { UserService } from './user.service';
+import { UserService } from '../services/users/user.service';
 
 @Injectable({
   providedIn: 'root'
 })
-export class LoginGuard implements CanActivate {
+export class RoleAuthGuard implements CanActivate {
   constructor(private auth: UserService, private router: Router) {}
   canActivate(
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
-      if (this.auth.isLoggedIn()) {
-        this.router.navigate(['index'], { replaceUrl: true });
-        return false;
+      const expectedRole = next.data.expectedRole;
+      const role = localStorage.getItem('role'); // #FIXME a changer avec un vrais service
+      if ( role.toString() !== expectedRole || !this.auth.isLoggedIn()) {
+        return false; // FIXME envoyer sur une autre route ?
       }
       return true;
   }
